Remove and update duplicate file
